refactor(ui-runtime): reorganize item aux and translation keys context
- Replace ItemAuxContext and TranslationKeysContext imports with new paths
- Remove obsolete ItemAuxData and ItemAuxMap files
- Introduce new ItemAux and TranslationKeys files for better structure
- Update references in related components and files
