+++ packages/ui-runtime/src/data/ItemAux.ts
@@ -1,5 +1,16 @@
 import { createContext } from '../core/fabric/context';
-import type { ItemAuxMap } from './ItemAuxMap';
+
+export interface ItemAuxData {
+  items: Record<string, number>;
+  correctionBoundaryAux: number;
+}
+
+/**
+ * Maps item type IDs (e.g. `"minecraft:apple"`) to their aux values (`raw_id * 65536`).
+ *
+ * Generated by the `item-aux` Regolith filter at `BP/scripts/data/itemAuxMap.generated.json`.
+ */
+export type ItemAuxMap = Record<string, number>;
 
 /**
  * Context that provides the item aux map (typeId → aux value) to the component tree.
