--- packages/ui-runtime/src/data/ItemAuxData.ts
@@ -1,4 +1,0 @@
-export interface ItemAuxData {
-  items: Record<string, number>;
-  correctionBoundaryAux: number;
-}
