--- packages/ui-runtime/src/data/ItemAuxMap.ts
@@ -1,18 +1,0 @@
-/**
- * Maps item type IDs (e.g. `"minecraft:apple"`) to their aux values (`raw_id * 65536`).
- *
- * Generated by the `item-aux` Regolith filter at `BP/scripts/data/itemAuxMap.generated.json`.
- * Provide it to the component tree via `ItemAuxContext`:
- *
- * ```tsx
- * import itemAuxMap from './data/itemAuxMap.generated.json';
- *
- * render(
- *   <ItemAuxContext value={itemAuxMap}>
- *     <MyInventory />
- *   </ItemAuxContext>,
- *   player,
- * );
- * ```
- */
-export type ItemAuxMap = Record<string, number>;
