+++ packages/ui-runtime/src/data/TranslationKeys.ts
@@ -1,5 +1,11 @@
 import { createContext } from '../core/fabric/context';
-import type { TranslationKeysMap } from './TranslationKeysMap';
+
+/**
+ * Maps translation keys (e.g. `"ui.myscreen.title"`) to their resolved display strings.
+ *
+ * Generated by the `translation-keys` Regolith filter at `BP/scripts/data/translationKeys.generated.json`.
+ */
+export type TranslationKeysMap = Record<string, string>;
 
 /**
  * Context that provides the translation keys map (key → resolved string) to the component tree.
