--- packages/ui-runtime/src/data/TranslationKeysMap.ts
@@ -1,18 +1,0 @@
-/**
- * Maps translation keys (e.g. `"ui.myscreen.title"`) to their resolved display strings.
- *
- * Generated by the `translation-keys` Regolith filter at `BP/scripts/data/translationKeys.generated.json`.
- * Provide it to the component tree via `TranslationKeysContext`:
- *
- * ```tsx
- * import translationKeys from './data/translationKeys.generated.json';
- *
- * render(
- *   <TranslationKeysContext value={translationKeys}>
- *     <MyScreen />
- *   </TranslationKeysContext>,
- *   player,
- * );
- * ```
- */
-export type TranslationKeysMap = Record<string, string>;
